+++ chitra/data_processing/default/__init__.py
@@ -1,2 +1,2 @@
-from .nlp import DefaultTextProcessor
-from .vision import DefaultVisionProcessor
+from chitra.data_processing.default.nlp import DefaultTextProcessor
+from chitra.data_processing.default.vision import DefaultVisionProcessor
