+++ chitra/image/image.py
