+++ chitra/image/tf_image.py
