+++ chitra/serve/__init__.py
@@ -1,3 +1,5 @@
 from chitra.serve.api import API, create_api
 from chitra.serve.app import GradioApp
 from chitra.serve.base import ModelServer
+from chitra.serve.cloud.aws_serverless import ChaliceServer
+from chitra.serve.cloud.base import CloudServer
