--- chitra/utility/__init__.py
@@ -1,3 +1,0 @@
-from .tf_utils import disable_gpu
-from .tf_utils import gpu_dynamic_mem_growth as tf_gpu_dynamic_mem_growth
-from .tf_utils import limit_gpu as tf_limit_gpu
