+++ tests/test_tf_image.py
@@ -3,7 +3,7 @@ import os
 import tensorflow as tf
 
 from chitra.image import Chitra
-from chitra.tf_image import read_image, resize_image
+from chitra.image.tf_image import read_image, resize_image
 
 chitra_banner = (
     "https://raw.githubusercontent.com/aniketmaurya/"
